cleanup and documentation
