+++ api-test/EditionTests.cs
@@ -5,9 +5,7 @@ using System.Threading.Tasks;
 using Bogus;
 using Dapper;
 using DeepEqual.Syntax;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.CodeAnalysis.Operations;
 using SQE.ApiTest.Helpers;
 using SQE.SqeApi.Server;
 using SQE.SqeApi.Server.DTOs;
