+++ data-access/EditionRepository.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data;
 using System.Threading.Tasks;
 using Dapper;
 using System.Linq;
-using System.Reflection.Metadata.Ecma335;
 using Microsoft.Extensions.Configuration;
 using System.Transactions;
 using SQE.SqeApi.DataAccess.Helpers;
