+++ data-access/Queries/Edition.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Text;
 
 namespace SQE.SqeApi.DataAccess.Queries
 {
