+++ sqe-api/DTOs/Artefact.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 
 namespace SQE.SqeApi.Server.DTOs
 {
