+++ sqe-api/DTOs/Editions.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using Newtonsoft.Json;
 
 namespace SQE.SqeApi.Server.DTOs
 {
