+++ sqe-api/DTOs/Sign.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using Microsoft.EntityFrameworkCore.ValueGeneration.Internal;
 
 namespace SQE.SqeApi.Server.DTOs
 {
