+++ sqe-api/Hubs/mainHubImagedObject.cs
@@ -1,28 +1,13 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.SignalR;
 using SQE.SqeApi.Server.DTOs;
-using SQE.SqeApi.Server.Services;
 
-namespace SQE.SqeApi.Server.Controllers
+namespace SQE.SqeApi.Server.Hubs
 {
-    [Authorize]
-    [ApiController]
-    public class ImagedObjectController : ControllerBase
+    public partial class MainHub : Hub
     {
-        private readonly IImagedObjectService _imagedObjectService;
-        private readonly IImageService _imageService;
-        private readonly IUserService _userService;
-
-        public ImagedObjectController(IImagedObjectService imagedObjectService, IImageService imageService,
-            IUserService userService)
-        {
-            _imagedObjectService = imagedObjectService;
-            _imageService = imageService;
-            _userService = userService;
-        }
-
         /// <summary>
         /// Provides information for the specified imaged object related to the specified edition, can include images and also their masks with optional.
         /// </summary>
@@ -30,9 +15,8 @@ namespace SQE.SqeApi.Server.Controllers
         /// <param name="imagedObjectId">Unique Id of the desired object from the imaging Institution</param>
         /// <param name="optional">Set 'artefacts' to receive related artefact data and 'masks' to include the artefact masks</param>
         [AllowAnonymous]
-        [HttpGet("v1/editions/{editionId}/imaged-objects/{imagedObjectId}")]
-        public async Task<ActionResult<ImagedObjectDTO>> GetEditionImagedObject([FromRoute] uint editionId,
-            [FromRoute] string imagedObjectId, [FromQuery] List<string> optional)
+        public async Task<ImagedObjectDTO> GetV1EditionsEditionIdImagedObjectsImagedObjectId(uint editionId,
+            string imagedObjectId, List<string> optional)
         {
             return await _imagedObjectService.GetImagedObjectAsync(
                 _userService.GetCurrentUserId(),
@@ -47,9 +31,8 @@ namespace SQE.SqeApi.Server.Controllers
         /// <param name="editionId">Unique Id of the desired edition</param>
         /// <param name="optional">Set 'artefacts' to receive related artefact data and 'masks' to include the artefact masks</param>
         [AllowAnonymous]
-        [HttpGet("v1/editions/{editionId}/imaged-objects")]
-        public async Task<ActionResult<ImagedObjectListDTO>> GetEditionImagedObjects([FromRoute] uint editionId,
-            [FromQuery] List<string> optional)
+        public async Task<ImagedObjectListDTO> GetV1EditionsEditionIdImagedObjects(uint editionId,
+            List<string> optional)
         {
             return await _imagedObjectService.GetImagedObjectsAsync(
                 _userService.GetCurrentUserId(),
@@ -61,8 +44,7 @@ namespace SQE.SqeApi.Server.Controllers
         /// Provides a list of all institutional image providers.
         /// </summary>
         [AllowAnonymous]
-        [HttpGet("v1/imaged-objects/institutions")]
-        public async Task<ActionResult<ImageInstitutionListDTO>> ListImageInstitutions()
+        public async Task<ImageInstitutionListDTO> GetV1ImagedObjectsInstitutions()
         {
             return await _imageService.GetImageInstitutionsAsync();
         }
