+++ sqe-api/Services/ImagedObjectService.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 using SQE.SqeApi.DataAccess;
